refactor: 출력 로직 개선
